Move components to root client directory
